Trim clap-utils/
