+++ clap-utils/Cargo.toml
@@ -11,7 +11,6 @@ edition = "2018"
 [dependencies]
 clap = "2.33.0"
 rpassword = "4.0"
-semver = "0.9.0"
 solana-remote-wallet = { path = "../remote-wallet", version = "0.24.0" }
 solana-sdk = { path = "../sdk", version = "0.24.0" }
 tiny-bip39 = "0.7.0"
