+++ install/Cargo.toml
@@ -29,6 +29,7 @@ solana-client = { path = "../client", version = "0.24.0" }
 solana-config-program = { path = "../programs/config", version = "0.24.0" }
 solana-logger = { path = "../logger", version = "0.24.0" }
 solana-sdk = { path = "../sdk", version = "0.24.0" }
+semver = "0.9.0"
 tar = "0.4.26"
 tempdir = "0.3.7"
 url = "2.1.1"
